Qt5 compat round 2: fix qHash signatures (size_t→QHashSeed), add qHash for ImageRequestKey, guard qt_add_translations, fix Qt5 CI env var, make LinguistTools optional

+++ src/Core/Flags64.hpp
@@ -69,7 +69,7 @@ template <typename Enum>
 }
 
 template <typename Enum>
-[[nodiscard]] inline size_t qHash(Flags64<Enum> flags, uint seed = 0) noexcept
+[[nodiscard]] inline uint qHash(Flags64<Enum> flags, uint seed = 0) noexcept
 {
     return qHash(flags.toInt(), seed);
 }
